Move translate methods to tim_translator.py
